+++ sdk/network/azure-mgmt-network/generated_samples/firewall_policy_rule_collection_group_with_http_headers_to_insert.py
@@ -66,6 +66,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-09-01/FirewallPolicyRuleCollectionGroupWithHttpHeadersToInsert.json
+# x-ms-original-file: 2026-01-01/FirewallPolicyRuleCollectionGroupWithHttpHeadersToInsert.json
 if __name__ == "__main__":
     main()