+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_virtual_hubs_operations_async.py
@@ -25,7 +25,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
         response = await self.client.virtual_hubs.get(
             resource_group_name=resource_group.name,
             virtual_hub_name="str",
-            api_version="2025-09-01",
+            api_version="2026-01-01",
         )
 
         # please add some check logic here by yourself
@@ -91,7 +91,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
                     "tags": {"str": "str"},
                     "type": "str",
                 },
-                api_version="2025-09-01",
+                api_version="2026-01-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -105,7 +105,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
             resource_group_name=resource_group.name,
             virtual_hub_name="str",
             virtual_hub_parameters={"tags": {"str": "str"}},
-            api_version="2025-09-01",
+            api_version="2026-01-01",
         )
 
         # please add some check logic here by yourself
@@ -118,7 +118,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
             await self.client.virtual_hubs.begin_delete(
                 resource_group_name=resource_group.name,
                 virtual_hub_name="str",
-                api_version="2025-09-01",
+                api_version="2026-01-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -130,7 +130,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
     async def test_virtual_hubs_list_by_resource_group(self, resource_group):
         response = self.client.virtual_hubs.list_by_resource_group(
             resource_group_name=resource_group.name,
-            api_version="2025-09-01",
+            api_version="2026-01-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself
@@ -140,7 +140,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
     @recorded_by_proxy_async
     async def test_virtual_hubs_list(self, resource_group):
         response = self.client.virtual_hubs.list(
-            api_version="2025-09-01",
+            api_version="2026-01-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself
@@ -153,7 +153,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
             await self.client.virtual_hubs.begin_get_effective_virtual_hub_routes(
                 resource_group_name=resource_group.name,
                 virtual_hub_name="str",
-                api_version="2025-09-01",
+                api_version="2026-01-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -168,7 +168,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
                 resource_group_name=resource_group.name,
                 virtual_hub_name="str",
                 get_inbound_routes_parameters={"connectionType": "str", "resourceUri": "str"},
-                api_version="2025-09-01",
+                api_version="2026-01-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -183,7 +183,7 @@ class TestNetworkManagementVirtualHubsOperationsAsync(AzureMgmtRecordedTestCase)
                 resource_group_name=resource_group.name,
                 virtual_hub_name="str",
                 get_outbound_routes_parameters={"connectionType": "str", "resourceUri": "str"},
-                api_version="2025-09-01",
+                api_version="2026-01-01",
             )
         ).result()  # call '.result()' to poll until service return final result